Create explicit rename API
Moving files had an undesired side-effect when renaming a file to an
already existing directory name. E.g. /file.txt to /dir if /dir is an
existing directory. This caused /file.txt to be moved to /dir but this
is undesired in cases where you really want to rename. This is why I
created an explicit API for renaming files. So move means moving things
to a directory now and rename is a 1:1 renaming of a file/directory to a
new name.